requirements.txt
torch==2.0.1
transformers==4.34.0
sentence-transformers==2.2.2
faiss-cpu==1.7.2
numpy==1.24.4
pandas==2.0.3
PyPDF2==3.0.1
pdfplumber==0.7.7
pytesseract==0.3.10
pdf2image==1.16.3
requests==2.30.0
nltk==3.8.1
google-generativeai==0.1.0
